requirements.txt
python-pptx
pypdf
pdfplumber
Pillow
